+++ src/graphql_clj/validator.clj
@@ -11,6 +11,7 @@
             [graphql-clj.validator.rules.fragments-on-composite-types]
             [graphql-clj.validator.rules.unique-variable-names]
             [graphql-clj.validator.rules.unique-operation-names]
+            [graphql-clj.validator.rules.unique-input-field-names]
             [graphql-clj.visitor :as visitor]
             [graphql-clj.spec :as spec]
             [instaparse.core :as insta]
@@ -23,6 +24,7 @@
             graphql-clj.validator.rules.known-argument-names/rules
             graphql-clj.validator.rules.known-fragment-names/rules
             graphql-clj.validator.rules.no-undefined-variables/rules
+            graphql-clj.validator.rules.unique-input-field-names/rules
             graphql-clj.validator.rules.arguments-of-correct-type/rules
             graphql-clj.validator.rules.default-values-of-correct-type/rules
             graphql-clj.validator.rules.variables-are-input-types/rules