+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file. This change
 - Validate that statements do not contain fragment cycles
 - Validate that fragments are on composite types
 - Make validation robust to order (can now reference types declared later in a statement or schema)
-- Validate unique variable and operation names
+- Validate unique variable, operation, and input type names
 
 ## [0.1.17] - 2016-10-14
 - Implement preliminary version of FieldsOnCorrectType validation